// gateway_service/src/generated/vacancy/v1/vacancy.ts
export enum VacancySearchField {
  VACANCY_SEARCH_FIELD_UNSPECIFIED = 0,
  VACANCY_SEARCH_FIELD_TITLE = 1,
  VACANCY_SEARCH_FIELD_DESCRIPTION = 2,
  VACANCY_SEARCH_FIELD_COMPANY_NAME = 3,
  UNRECOGNIZED = -1,
}
// ...
export interface CreateVacancyRequest {
  title: string;
  description: string;
  salary: number;
  link: string;
  city: string;
  company_name: string;
}

export interface CreateVacancyResponse {
  vacancy?: Vacancy | undefined;
}

export interface BatchCreateVacanciesRequest {
  vacancies: CreateVacancyRequest[];
}
// ...
export const JOBAGGREGATOR_VACANCY_V1_PACKAGE_NAME = "jobaggregator.vacancy.v1";